refactor: summary/post page, component 분리
